Fix compile errors: sealed trait compliance, init order, proto serde, exhaustive matches
- Move TemplateBoundPartyMapping into TopologyMapping.scala (sealed trait)
- Add templateBoundPartyConfig to Forwarding/Caching/ValidatingTopologySnapshot
- Add TemplateBoundPartyMapping case to TopologyStateProcessor exhaustive match
- Fix companion object init: val→def for code (Scala super-constructor ordering)
- Fix fromProtoV30: use participantIdFromProtoPrimitive for UID-format proto field
- Fix TemplateBoundPartyExtractor: use contract.templateId and contract.signatories
- Fix TemplateBoundPartyChecks rejection type to InvalidTopologyMapping
- Fix tests: toProtoV30→toProto, restrictedToDomain→restrictedToSynchronizer
- Remove unused imports across all TBP files

All 26 community-base TBP tests pass.
